usc: init controller's caches before syncing
